<Refactor> [Menu Handler]: Moved menu handler to correct folder and added a namespace
